+++ test/common.go
@@ -2,12 +2,7 @@
 package test
 
 import (
-	"net/url"
-	"testing"
-
 	"github.com/satori/go.uuid"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 // CreateRandomValidTestName functions creates a valid length name
@@ -17,17 +12,4 @@ func CreateRandomValidTestName(name string) string {
 		return randomName[:61]
 	}
 	return randomName
-}
-
-func EqualURLs(t *testing.T, expected string, actual string) {
-	expectedURL, err := url.Parse(expected)
-	require.Nil(t, err)
-	actualURL, err := url.Parse(actual)
-	require.Nil(t, err)
-	assert.Equal(t, expectedURL.Scheme, actualURL.Scheme)
-	assert.Equal(t, expectedURL.Host, actualURL.Host)
-	assert.Equal(t, len(expectedURL.Query()), len(actualURL.Query()))
-	for name, value := range expectedURL.Query() {
-		assert.Equal(t, value, actualURL.Query()[name])
-	}
 }
